% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocklist-package.r
\docType{package}
\name{blocklist}
\alias{blocklist}
\alias{blocklist-package}
\title{Tools to Work with the <blocklist.de> 'API'}
\description{
Tools to Work with the <blocklist.de> 'API'
}
\author{
Bob Rudis (@hrbrmstr)
}

