% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocklist.r
\name{get_last_added_ips}
\alias{get_last_added_ips}
\title{Query <blocklist.de> API for IPs blocked in since a period in time, optionally
filtering by service}
\usage{
get_last_added_ips(since = NULL, service = c("all", "amavis", "apacheddos",
  "asterisk", "badbot", "ftp", "imap", "ircbot", "mail", "pop3", "regbot",
  "rfi-attack", "sasl", "ssh", "w00tw00t", "portflood", "sql-injection",
  "webmin", "trigger-spam", "manuall", "bruteforcelogin"))
}
\arguments{
\item{since}{either a UNIX timestamp (e.g. \code{1459736686}), a string in "\code{HH:MM}"
format or a time difference in seconds (e..g \code{3600}). The API will
return all requested IPs logged since that time. Leave \code{NULL} for the
API default.}

\item{service}{only return addresses for a given service. Default: \code{all} services.}
}
\description{
Query <blocklist.de> API for IPs blocked in since a period in time, optionally
filtering by service
}
\examples{
# defaults
get_last_added_ips()

# last hour
get_last_added_ips(3600)

# since 3PM today
get_last_added_ips("15:00")

# for ssh blocks in the last hour
get_last_added_ips(3600, "ssh")
}

